+++ R/amerasfit-class.R
@@ -105,15 +105,15 @@ summary.amerasfit <- function(object, ...) {
 
       if (object$CI.computed) {
         CI <- y$CI
-        CI.lowerbound <- CI.upperbound <- coef * NA
-        CI.lowerbound[match(rownames(CI), names(coef))] <- CI$lower
-        CI.upperbound[match(rownames(CI), names(coef))] <- CI$upper
+        CI.lower <- CI.upper <- coef * NA
+        CI.lower[match(rownames(CI), names(coef))] <- CI$lower
+        CI.upper[match(rownames(CI), names(coef))] <- CI$upper
 
         res <- cbind(
           res,
           data.frame(
-            CI.lowerbound = CI.lowerbound,
-            CI.upperbound = CI.upperbound
+            CI.lower = CI.lower,
+            CI.upper = CI.upper
           )
         )
         # Add p-value columns for profile likelihood intervals. If not used, these will be removed after